+++ spec/spec_helper.rb
@@ -44,10 +44,12 @@ end
 require "wild"
 require_relative "support/wild_hooks/hook_fixtures"
 require_relative "support/wild_skillops/fixtures"
+require_relative "support/wild_analyzers_permission/fixtures"
 
 RSpec.configure do |config|
   config.include Wild::Hooks::TestSupport::HookFixtures
   config.include Wild::Skillops::TestSupport::Fixtures
+  config.include Wild::Analyzers::Permission::TestSupport::Fixtures
   config.expect_with :rspec do |expectations|
     expectations.include_chain_clauses_in_custom_matcher_descriptions = true
   end